requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "3.13"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "3.13"
cachetools==5.5.2 ; python_version >= "3.11" and python_version < "3.13"
certifi==2025.4.26 ; python_version >= "3.11" and python_version < "3.13"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_version < "3.13"
click==8.1.8 ; python_version >= "3.11" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "3.13" and platform_system == "Windows"
fastapi==0.115.12 ; python_version >= "3.11" and python_version < "3.13"
google-api-core==2.25.1 ; python_version >= "3.11" and python_version < "3.13"
google-auth==2.40.3 ; python_version >= "3.11" and python_version < "3.13"
google-cloud-vision==3.10.2 ; python_version >= "3.11" and python_version < "3.13"
google-genai==1.26.0 ; python_version >= "3.11" and python_version < "3.13"
googleapis-common-protos==1.70.0 ; python_version >= "3.11" and python_version < "3.13"
grpcio-status==1.74.0 ; python_version >= "3.11" and python_version < "3.13"
grpcio==1.74.0 ; python_version >= "3.11" and python_version < "3.13"
h11==0.16.0 ; python_version >= "3.11" and python_version < "3.13"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "3.13"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "3.13"
idna==3.10 ; python_version >= "3.11" and python_version < "3.13"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "3.13"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "3.13"
networkx==3.4.2 ; python_version >= "3.11" and python_version < "3.13"
numpy==1.26.4 ; python_version >= "3.11" and python_version < "3.13"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "3.13"
proto-plus==1.26.1 ; python_version >= "3.11" and python_version < "3.13"
protobuf==6.32.0 ; python_version >= "3.11" and python_version < "3.13"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "3.13"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "3.13"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "3.13"
pydantic==2.11.4 ; python_version >= "3.11" and python_version < "3.13"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "3.13"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "3.13"
python-multipart==0.0.20 ; python_version >= "3.11" and python_version < "3.13"
pytz==2025.2 ; python_version >= "3.11" and python_version < "3.13"
requests==2.31.0 ; python_version >= "3.11" and python_version < "3.13"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "3.13"
six==1.17.0 ; python_version >= "3.11" and python_version < "3.13"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "3.13"
starlette==0.46.2 ; python_version >= "3.11" and python_version < "3.13"
tenacity==8.5.0 ; python_version >= "3.11" and python_version < "3.13"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_version < "3.13"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_version < "3.13"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "3.13"
urllib3==2.4.0 ; python_version >= "3.11" and python_version < "3.13"
uvicorn==0.34.2 ; python_version >= "3.11" and python_version < "3.13"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "3.13"
